getNodesBy options and module
