+++ src/nodejs/neo-rpc/v2.9.0/client/cli/getNodesBy.js
@@ -8,12 +8,12 @@ const program = require('commander')
 const _       = require('underscore')
 
 const dbg     = require('nodejs_util/debug')
-const netutil = require('nodejs_util/network')
+const netUtil = require('nodejs_util/network')
 
 var cfg       = require('nodejs_config/config.js')
 var config    = cfg.load('nodejs_config/nodejs.config.json')
 
-const command = require('nodejs_neo-rpc/v2.9.0/client/module/neo-rpc')
+const getNodesBy = require('nodejs_neo-rpc/v2.9.0/client/module/getNodesBy')
 
 
 let nodes = []
@@ -30,25 +30,27 @@ program
   .option('-d, --debug', 'Debug')
   .option('-n, --node [node]', 'Set RPC node to use (be sure to preface with https://), if not provided will try to use node with tallest block')
   .option('-N, --Net [Net]', 'Select network [net]: i.e., TestNet or MainNet', 'TestNet')
-  .option('-m, --method [method]', 'Get nodes by the given criteria, default is ping', 'getNodesByPing')
+  .option('-m, --method [method]', 'Get nodes by the given criteria, default is ping', 'ping')
+  .option('-o, --order [order]', 'Order by \'asc\' (ascending <- default) or \'dsc\' (descending)', 'asc')
   .on('--help', function(){
-    print('Note: -m --method options are: "getNodesByPing", "getNodesByTallest", "getNodesByConnections", getNodesByVersion')
+    print('Note: -m --method options are: "ping", "tallest", "connection", "version", "rawMemPool"')
   })
   .parse(process.argv)
 
 if (program.debug) {
   print('DEBUGGING: ' + __filename)
   defly = true
-  netutil.debug()
+  netUtil.debug()
 }
 
 let options = {
-  net: program.Net
+  net: program.Net,
+  order: program.order
 }
 
 if (program.node) options.nodes = [{ 'url': program.node }]
 
-netutil[program.method](options).then(rankedNodes => {
+getNodesBy[program.method.toLowerCase()](options).then(rankedNodes => {
 
   if (defly) dbg.logDeep(__filename + ': getNodesByPing().rankedNodes: ', rankedNodes)
   nodes = rankedNodes
